requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
appdirs==1.4.4
bullet-safety-gym==1.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
docker-pycreds==0.4.0
dsrl==0.1.0
easy-runner==0.1.0
Farama-Notifications==0.0.4
fast-safe-rl==0.1.0
filelock==3.17.0
fsrl @ git+https://github.com/liuzuxin/fsrl.git@e056fc9498d5d037869533da7cf976acf462f918
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
glfw==2.8.0
grpcio==1.71.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.28.1
h5py==3.13.0
idna==3.10
imageio==2.37.0
jax-jumpy==1.0.0
Jinja2==3.1.6
llvmlite==0.44.0
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
mujoco==2.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
osrl-lib==0.1.0
packaging==24.2
pathtools==0.1.2
pettingzoo==1.24.3
pillow==11.1.0
platformdirs==4.3.6
prettytable==3.7.0
protobuf==3.19.6
psutil==5.9.8
pybullet==3.2.7
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.1.0
PyOpenGL==3.1.9
pyrallis==0.3.1
python-dateutil==2.8.2
PyYAML==6.0.2
requests==2.32.3
safety-gymnasium==0.4.0
scipy==1.10.1
sentry-sdk==2.22.0
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
swig==4.1.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tianshou==0.5.1
torch==1.13.1
tqdm==4.67.1
triton==3.2.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
wandb==0.14.2
wcwidth==0.2.13
Werkzeug==3.1.3
xmltodict==0.14.2
